refactor(repl): split command handling from REPL model
- move slash-command dispatch and command logic into command_handlers.go
- move REPL utility and helper methods into repl_helpers.go
- relocate command-focused tests into command_handlers_test.go
- preserve completed stream segments when retrying failed responses
- reset session-scoped tool permissions when starting a new session
- update README demo asset
- add issue 29 planning notes
